Expand git searches
to look for the upstream commit hash and
the CVE identifier when looking to see if
a patch has already been applied.

CVE IDs aren't often present in commit messages,
but occasionally they are, so catch them.

And often when patches are applied to stable
branches, the upstream commit hash is mentioned,
so looking for that helps cut down on false
negatives.